requirements.txt
yfinance
pandas
numpy
scikit-learn
matplotlib
PyYAML
